bot/requirements.txt
discord.py
dotenv
requests